auth-service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pymongo>=4.3.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.24.0
cloudinary>=1.33.0
python-dotenv>=1.0.0
pydantic>=1.10.7